using decimal for price

+++ inventory/models.py
@@ -13,7 +13,7 @@ class ItemCategory(models.Model):
 class Item(models.Model):
     barcode = models.CharField(max_length=30, unique=True)
     name = models.CharField(max_length=100)
-    unit_price = models.FloatField()
+    unit_price = models.DecimalField(max_digits=20, decimal_places=2)
     category = models.ForeignKey(ItemCategory, on_delete=models.RESTRICT)
     notes = models.TextField(blank=True)
